Simplfied using Redux and enforced type safety
