+++ src/redux/actionTypes.ts
@@ -1,3 +1,11 @@
 export enum UserActionTypes {
     SET_USER = "@User/SET_USER"
 }
+
+export interface setUserAction {
+    type: typeof UserActionTypes.SET_USER
+    payload: User
+}
+
+// The collective type for all actions converning the user
+export type UserActionType = setUserAction
