+++ src/redux/index.ts
@@ -6,4 +6,6 @@ const rootReducer = combineReducers({
     user: userReducer,
 });
 
+export type AppState = ReturnType<typeof rootReducer>
+
 export default rootReducer;
