+++ src/redux/store.ts
@@ -1,7 +1,9 @@
+import { AppState } from "@auth0/auth0-react";
 import { createStore, applyMiddleware, compose } from "redux"
-import thunk from "redux-thunk"
+import thunk, { ThunkMiddleware } from "redux-thunk"
 
 import rootReducer from "./index"
+import { AppActionType } from "./state";
 
 declare global {
     interface Window {
@@ -12,5 +14,5 @@ declare global {
 const composeEnhancers = (typeof window !== 'undefined' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose;
 
 export const store = createStore(rootReducer, /* preloadedState, */ composeEnhancers(
-    applyMiddleware(thunk))
+    applyMiddleware(thunk as ThunkMiddleware<AppState, AppActionType>))
 );
