+++ src/types/User.d.ts
@@ -1,0 +1,12 @@
+type User = {
+    email: string
+    email_verified: boolean
+    family_name: string
+    given_name: string
+    locale: string
+    name: string
+    nickname: string
+    picture: string
+    sub: string
+    updated_at: string
+}
